fix(wallet-adapter): update test for wallet adpaters
