+++ packages/libs/wallet-adapter-magic/vitest.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     globals: true,
-    environment: 'node',
+    environment: 'jsdom',
     include: ['src/__tests__/**/*.{test,spec}.{ts,tsx}'],
   },
 });
